requirements.txt
astroid==2.1.0
autopep8==1.4.3
boto3==1.9.88
botocore==1.12.88
CacheControl==0.12.5
certifi==2018.11.29
chardet==3.0.4
Django==2.1.5
django-crispy-forms==1.7.2
django-hosts==3.0
django-widget-tweaks==1.4.3
docutils==0.14
idna==2.8
iso3166==0.9
isort==4.3.4
jmespath==0.9.3
lazy-object-proxy==1.3.1
mapbox==0.17.2
mccabe==0.6.1
msgpack==0.6.1
polyline==1.3.2
pycodestyle==2.5.0
pylint==2.2.2
pytemperature==1.0
python-dateutil==2.8.0
pytz==2018.9
requests==2.21.0
s3transfer==0.1.13
simplejson==3.16.0
six==1.12.0
uritemplate==3.0.0
urllib3==1.24.1
wrapt==1.11.1
